backend/requirements.txt
flask==3.0.0
requests==2.31.0
pandas==2.1.3
ta==0.11.0
openai>=1.12.0
websocket-client==1.6.4
python-dotenv==1.0.0
flask-cors==4.0.0
schedule==1.2.0
gunicorn==21.2.0
matplotlib==3.8.0

